requirements.txt
selenium
python-vlc
requests
webdriver-manager
